Make ID2D1Shader.GetInputType branchless

+++ src/ComputeSharp.D2D1/Interfaces/__Internals/ID2D1Shader.cs
@@ -25,7 +25,7 @@ public interface ID2D1Shader
     /// <returns>The number of inputs for the current shader.</returns>
     [EditorBrowsable(EditorBrowsableState.Never)]
     [Obsolete("This method is not intended to be used directly by user code")]
-    byte GetInputCount();
+    uint GetInputCount();
 
     /// <summary>
     /// Gets the input type for the input at the input index.
@@ -35,7 +35,7 @@ public interface ID2D1Shader
     /// <remarks>The return value if the input is out of range is undefined.</remarks>
     [EditorBrowsable(EditorBrowsableState.Never)]
     [Obsolete("This method is not intended to be used directly by user code")]
-    byte GetInputType(byte index);
+    uint GetInputType(uint index);
 
     /// <summary>
     /// Loads the dispatch data for the shader.
